feat: add getRasterFunction4DataAggregation to retrieve raster function for SIHUI task
