--- src/surface-temp/services/SIHUI/DataAggregation/RasterFunctionTemplate4DataAggregation.ts
@@ -1,960 +1,0 @@
-const RASTER_FUNCTION_TEMPLATE_DATA_AGGREGATION = {
-    name: 'GenerateRaster',
-    parameters: {
-        rasterFunction: {
-            description:
-                'This RFT creates a cloud free mean composite of the Landsat Level-2 bands required for the the Landsat SIUHI worfklow.',
-            name: 'Landsat Level-2 SIUHI - Part 1 of 3 - Data Aggregation',
-            arguments: {
-                type: 'RasterCollectionFunctionArguments',
-                RasterCollection: {
-                    name: 'RasterCollection',
-                    isDataset: true,
-                    isPublic: true,
-                    type: 'RasterFunctionVariable',
-                    _object_id: 3,
-                    value: {
-                        // "url": "https://landsat.imagery1.arcgis.com/arcgis/rest/services/LandsatC2L2/ImageServer?token={{token}}",
-                        url: '{{landsatImageryServiceUrl}}?token={{token}}',
-                        name: 'Landsat Level-2',
-                        renderingRule: {
-                            rasterFunction: 'None',
-                        },
-                        mosaicRule: {
-                            ascending: true,
-                            mosaicMethod: 'esriMosaicAttribute',
-                            mosaicOperation: 'MT_FIRST',
-                            sortField: 'best',
-                            sortValue: '0',
-                        },
-                    },
-                },
-                QueryGeometry: {
-                    name: 'QueryGeometry',
-                    isDataset: false,
-                    isPublic: true,
-                    type: 'RasterFunctionVariable',
-                    _object_id: 4,
-                    value: {
-                        // "geometries": [
-                        //   {
-                        //     "rings": [
-                        //       [
-                        //         [
-                        //           -11162901.410114916,
-                        //           2949214.897361696
-                        //         ],
-                        //         [
-                        //           -11163720.942599999,
-                        //           2950961.5565000027
-                        //         ],
-                        //         [
-                        //           -11164905.0454,
-                        //           2950961.5565000027
-                        //         ],
-                        //         [
-                        //           -11165392.6786,
-                        //           2952000.1812000014
-                        //         ],
-                        //         [
-                        //           -11166576.8331,
-                        //           2952000.1812000014
-                        //         ],
-                        //         [
-                        //           -11167064.7379,
-                        //           2953038.9069
-                        //         ],
-                        //         [
-                        //           -11174169.9753,
-                        //           2953038.9069
-                        //         ],
-                        //         [
-                        //           -11174658.410799999,
-                        //           2954077.733600002
-                        //         ],
-                        //         [
-                        //           -11175842.6687,
-                        //           2954077.733600002
-                        //         ],
-                        //         [
-                        //           -11176820.3038,
-                        //           2956155.6900000013
-                        //         ],
-                        //         [
-                        //           -11178004.6654,
-                        //           2956155.6900000013
-                        //         ],
-                        //         [
-                        //           -11178493.8654,
-                        //           2957194.8197000027
-                        //         ],
-                        //         [
-                        //           -11182047.1056,
-                        //           2957194.8197000027
-                        //         ],
-                        //         [
-                        //           -11181557.7502,
-                        //           2956155.6900000013
-                        //         ],
-                        //         [
-                        //           -11182742.1118,
-                        //           2956155.6900000013
-                        //         ],
-                        //         [
-                        //           -11183231.519,
-                        //           2957194.8197000027
-                        //         ],
-                        //         [
-                        //           -11184415.9325,
-                        //           2957194.8197000027
-                        //         ],
-                        //         [
-                        //           -11184905.6122,
-                        //           2958234.0507000014
-                        //         ],
-                        //         [
-                        //           -11187274.5428,
-                        //           2958234.0507000014
-                        //         ],
-                        //         [
-                        //           -11190217.881099999,
-                        //           2964471.563500002
-                        //         ],
-                        //         [
-                        //           -11189033.1042,
-                        //           2964471.563500002
-                        //         ],
-                        //         [
-                        //           -11188051.2115,
-                        //           2962391.9869999997
-                        //         ],
-                        //         [
-                        //           -11186866.5386,
-                        //           2962391.9869999997
-                        //         ],
-                        //         [
-                        //           -11186375.9757,
-                        //           2961352.351
-                        //         ],
-                        //         [
-                        //           -11184006.7337,
-                        //           2961352.351
-                        //         ],
-                        //         [
-                        //           -11184497.1927,
-                        //           2962391.9869999997
-                        //         ],
-                        //         [
-                        //           -11183312.5198,
-                        //           2962391.9869999997
-                        //         ],
-                        //         [
-                        //           -11183803.147599999,
-                        //           2963431.7245000005
-                        //         ],
-                        //         [
-                        //           -11182618.422799999,
-                        //           2963431.7245000005
-                        //         ],
-                        //         [
-                        //           -11182127.8468,
-                        //           2962391.9869999997
-                        //         ],
-                        //         [
-                        //           -11179758.501,
-                        //           2962391.9869999997
-                        //         ],
-                        //         [
-                        //           -11179268.249699999,
-                        //           2961352.351
-                        //         ],
-                        //         [
-                        //           -11178083.6288,
-                        //           2961352.351
-                        //         ],
-                        //         [
-                        //           -11177593.6503,
-                        //           2960312.816200003
-                        //         ],
-                        //         [
-                        //           -11174039.9432,
-                        //           2960312.816200003
-                        //         ],
-                        //         [
-                        //           -11174529.7658,
-                        //           2961352.351
-                        //         ],
-                        //         [
-                        //           -11173345.1448,
-                        //           2961352.351
-                        //         ],
-                        //         [
-                        //           -11174325.3484,
-                        //           2963431.7245000005
-                        //         ],
-                        //         [
-                        //           -11175510.0733,
-                        //           2963431.7245000005
-                        //         ],
-                        //         [
-                        //           -11176000.5582,
-                        //           2964471.563500002
-                        //         ],
-                        //         [
-                        //           -11177185.335099999,
-                        //           2964471.563500002
-                        //         ],
-                        //         [
-                        //           -11179641.335199999,
-                        //           2969672.281800002
-                        //         ],
-                        //         [
-                        //           -11180826.372399999,
-                        //           2969672.281800002
-                        //         ],
-                        //         [
-                        //           -11181810.4253,
-                        //           2971753.281100001
-                        //         ],
-                        //         [
-                        //           -11180625.2837,
-                        //           2971753.281100001
-                        //         ],
-                        //         [
-                        //           -11181950.59199256,
-                        //           2974553.772238198
-                        //         ],
-                        //         [
-                        //           -11153503.769933363,
-                        //           2974553.772238198
-                        //         ],
-                        //         [
-                        //           -11153164.215699999,
-                        //           2973834.6878000014
-                        //         ],
-                        //         [
-                        //           -11151978.969899999,
-                        //           2973834.6878000014
-                        //         ],
-                        //         [
-                        //           -11150996.745199999,
-                        //           2971753.281100001
-                        //         ],
-                        //         [
-                        //           -11149811.603699999,
-                        //           2971753.281100001
-                        //         ],
-                        //         [
-                        //           -11149320.8749,
-                        //           2970712.7305000015
-                        //         ],
-                        //         [
-                        //           -11146950.6962,
-                        //           2970712.7305000015
-                        //         ],
-                        //         [
-                        //           -11145970.109299999,
-                        //           2968631.9348000027
-                        //         ],
-                        //         [
-                        //           -11147155.0945,
-                        //           2968631.9348000027
-                        //         ],
-                        //         [
-                        //           -11146175.285699999,
-                        //           2966551.545900002
-                        //         ],
-                        //         [
-                        //           -11147360.1667,
-                        //           2966551.545900002
-                        //         ],
-                        //         [
-                        //           -11145402.9858,
-                        //           2962391.9869999997
-                        //         ],
-                        //         [
-                        //           -11146587.6587,
-                        //           2962391.9869999997
-                        //         ],
-                        //         [
-                        //           -11145121.9292,
-                        //           2959273.3828000017
-                        //         ],
-                        //         [
-                        //           -11143937.4121,
-                        //           2959273.3828000017
-                        //         ],
-                        //         [
-                        //           -11143449.327499999,
-                        //           2958234.0507000014
-                        //         ],
-                        //         [
-                        //           -11142264.8622,
-                        //           2958234.0507000014
-                        //         ],
-                        //         [
-                        //           -11141777.0494,
-                        //           2957194.8197000027
-                        //         ],
-                        //         [
-                        //           -11140592.6359,
-                        //           2957194.8197000027
-                        //         ],
-                        //         [
-                        //           -11139617.773,
-                        //           2955116.661200002
-                        //         ],
-                        //         [
-                        //           -11140802.082799999,
-                        //           2955116.661200002
-                        //         ],
-                        //         [
-                        //           -11140314.929,
-                        //           2954077.733600002
-                        //         ],
-                        //         [
-                        //           -11141499.1869,
-                        //           2954077.733600002
-                        //         ],
-                        //         [
-                        //           -11141012.2009,
-                        //           2953038.9069
-                        //         ],
-                        //         [
-                        //           -11142196.4071,
-                        //           2953038.9069
-                        //         ],
-                        //         [
-                        //           -11141709.5889,
-                        //           2952000.1812000014
-                        //         ],
-                        //         [
-                        //           -11142893.7433,
-                        //           2952000.1812000014
-                        //         ],
-                        //         [
-                        //           -11143380.6133,
-                        //           2953038.9069
-                        //         ],
-                        //         [
-                        //           -11145749.025799999,
-                        //           2953038.9069
-                        //         ],
-                        //         [
-                        //           -11145262.052299999,
-                        //           2952000.1812000014
-                        //         ],
-                        //         [
-                        //           -11146446.206799999,
-                        //           2952000.1812000014
-                        //         ],
-                        //         [
-                        //           -11145959.400999999,
-                        //           2950961.5565000027
-                        //         ],
-                        //         [
-                        //           -11147143.5038,
-                        //           2950961.5565000027
-                        //         ],
-                        //         [
-                        //           -11148327.6066,
-                        //           2950961.5565000027
-                        //         ],
-                        //         [
-                        //           -11147840.9169,
-                        //           2949923.0326000005
-                        //         ],
-                        //         [
-                        //           -11151393.0701,
-                        //           2949923.0326000005
-                        //         ],
-                        //         [
-                        //           -11151879.9149,
-                        //           2950961.5565000027
-                        //         ],
-                        //         [
-                        //           -11153064.0176,
-                        //           2950961.5565000027
-                        //         ],
-                        //         [
-                        //           -11154526.024799999,
-                        //           2954077.733600002
-                        //         ],
-                        //         [
-                        //           -11155710.2829,
-                        //           2954077.733600002
-                        //         ],
-                        //         [
-                        //           -11156198.1099,
-                        //           2955116.661200002
-                        //         ],
-                        //         [
-                        //           -11157382.4197,
-                        //           2955116.661200002
-                        //         ],
-                        //         [
-                        //           -11157870.5184,
-                        //           2956155.6900000013
-                        //         ],
-                        //         [
-                        //           -11160239.2416,
-                        //           2956155.6900000013
-                        //         ],
-                        //         [
-                        //           -11159751.0392,
-                        //           2955116.661200002
-                        //         ],
-                        //         [
-                        //           -11160935.349,
-                        //           2955116.661200002
-                        //         ],
-                        //         [
-                        //           -11160447.3148,
-                        //           2954077.733600002
-                        //         ],
-                        //         [
-                        //           -11161631.5728,
-                        //           2954077.733600002
-                        //         ],
-                        //         [
-                        //           -11159349.444864925,
-                        //           2949214.897361696
-                        //         ],
-                        //         [
-                        //           -11162901.410114916,
-                        //           2949214.897361696
-                        //         ]
-                        //       ]
-                        //     ]
-                        //   }
-                        // ],
-                        geometries: '{{queryGeometry}}',
-                        spatialReference: {
-                            latestWkid: 3857,
-                            wkid: 102100,
-                        },
-                    },
-                },
-                WhereClause: {
-                    name: 'WhereClause',
-                    isDataset: false,
-                    isPublic: true,
-                    type: 'RasterFunctionVariable',
-                    _object_id: 5,
-                    value: 'year = {{year}} AND month IN ({{months}})',
-                },
-                AggregationDefinition: {
-                    name: 'AggregationDefinition',
-                    value: {
-                        definitionType: 'ALL',
-                        dimension: 'acquisitiondate',
-                    },
-                    isDataset: false,
-                    isPublic: true,
-                    type: 'RasterFunctionVariable',
-                    _object_id: 6,
-                },
-                ItemFunction: {
-                    description: 'lscft',
-                    name: 'lscft',
-                    arguments: {
-                        type: 'ClipFunctionArguments',
-                        Raster: {
-                            description: 'A raster function template.',
-                            name: 'Raster Function Template',
-                            arguments: {
-                                type: 'ExtractBandFunctionArguments',
-                                Raster: {
-                                    name: 'Raster',
-                                    isDataset: true,
-                                    isPublic: true,
-                                    type: 'RasterFunctionVariable',
-                                    _object_id: 11,
-                                    value: {
-                                        url: '{{landsatImageryServiceUrl}}?token={{token}}',
-                                        name: 'Landsat Level-2',
-                                        renderingRule: {
-                                            rasterFunction: 'None',
-                                        },
-                                        mosaicRule: {
-                                            ascending: true,
-                                            mosaicMethod: 'esriMosaicAttribute',
-                                            mosaicOperation: 'MT_FIRST',
-                                            sortField: 'best',
-                                            sortValue: '0',
-                                        },
-                                    },
-                                },
-                                BandIDs: {
-                                    name: 'BandIds4',
-                                    value: [2, 3, 4, 5, 8],
-                                    isDataset: false,
-                                    isPublic: false,
-                                    type: 'RasterFunctionVariable',
-                                    _object_id: 12,
-                                },
-                                BandNames: {
-                                    name: 'BandNames4',
-                                    value: ['Red', 'Green', 'Blue'],
-                                    isDataset: false,
-                                    isPublic: false,
-                                    type: 'RasterFunctionVariable',
-                                    _object_id: 14,
-                                },
-                                BandWavelengths: {
-                                    name: 'BandWavelengths4',
-                                    value: [600, 400, 200],
-                                    isDataset: false,
-                                    isPublic: false,
-                                    type: 'RasterFunctionVariable',
-                                    _object_id: 16,
-                                },
-                                Method: {
-                                    name: 'Method4',
-                                    value: 2,
-                                    isDataset: false,
-                                    isPublic: false,
-                                    type: 'RasterFunctionVariable',
-                                    _object_id: 18,
-                                },
-                                MissingBandAction: {
-                                    name: 'MissingBandAction4',
-                                    value: 0,
-                                    isDataset: false,
-                                    isPublic: false,
-                                    type: 'RasterFunctionVariable',
-                                    _object_id: 19,
-                                },
-                                _object_id: 10,
-                            },
-                            function: {
-                                pixelType: 'UNKNOWN',
-                                name: 'Extract Green Red NIR SWIR1 TIRS',
-                                description:
-                                    'Specifies which bands to work with when using multiband datasets.',
-                                type: 'ExtractBandFunction',
-                                _object_id: 20,
-                            },
-                            functionType: 0,
-                            properties: {
-                                MatchVariable: {
-                                    name: 'MatchVariable',
-                                    value: true,
-                                    isDataset: false,
-                                    isPublic: false,
-                                    type: 'RasterFunctionVariable',
-                                    _object_id: 22,
-                                },
-                                UnionDimension: {
-                                    name: 'UnionDimension',
-                                    value: false,
-                                    isDataset: false,
-                                    isPublic: false,
-                                    type: 'RasterFunctionVariable',
-                                    _object_id: 23,
-                                },
-                                type: 'PropertySet',
-                                _object_id: 21,
-                            },
-                            type: 'RasterFunctionTemplate',
-                            _object_id: 9,
-                        },
-                        ClippingRaster: {
-                            description: 'A raster function template.',
-                            name: 'Raster Function Template',
-                            arguments: {
-                                type: 'TransposeBitsFunctionArguments',
-                                QualityClass: 'qcbp_UserDefined',
-                                Raster: {
-                                    description: 'A raster function template.',
-                                    name: 'Raster Function Template',
-                                    arguments: {
-                                        type: 'ExtractBandFunctionArguments',
-                                        Raster: {
-                                            _object_ref_id: 11,
-                                        },
-                                        BandIDs: {
-                                            name: 'BandIds5',
-                                            value: [7],
-                                            isDataset: false,
-                                            isPublic: false,
-                                            type: 'RasterFunctionVariable',
-                                            _object_id: 29,
-                                        },
-                                        BandNames: {
-                                            name: 'BandNames5',
-                                            value: ['Red', 'Green', 'Blue'],
-                                            isDataset: false,
-                                            isPublic: false,
-                                            type: 'RasterFunctionVariable',
-                                            _object_id: 31,
-                                        },
-                                        BandWavelengths: {
-                                            name: 'BandWavelengths5',
-                                            value: [600, 400, 200],
-                                            isDataset: false,
-                                            isPublic: false,
-                                            type: 'RasterFunctionVariable',
-                                            _object_id: 33,
-                                        },
-                                        Method: {
-                                            name: 'Method5',
-                                            value: 2,
-                                            isDataset: false,
-                                            isPublic: false,
-                                            type: 'RasterFunctionVariable',
-                                            _object_id: 35,
-                                        },
-                                        MissingBandAction: {
-                                            name: 'MissingBandAction5',
-                                            value: 0,
-                                            isDataset: false,
-                                            isPublic: false,
-                                            type: 'RasterFunctionVariable',
-                                            _object_id: 36,
-                                        },
-                                        _object_id: 27,
-                                    },
-                                    function: {
-                                        pixelType: 'U16',
-                                        name: 'Extract QA Band',
-                                        description:
-                                            'Specifies which bands to work with when using multiband datasets.',
-                                        type: 'ExtractBandFunction',
-                                        _object_id: 37,
-                                    },
-                                    functionType: 0,
-                                    properties: {
-                                        MatchVariable: {
-                                            name: 'MatchVariable',
-                                            value: true,
-                                            isDataset: false,
-                                            isPublic: false,
-                                            type: 'RasterFunctionVariable',
-                                            _object_id: 39,
-                                        },
-                                        UnionDimension: {
-                                            name: 'UnionDimension',
-                                            value: false,
-                                            isDataset: false,
-                                            isPublic: false,
-                                            type: 'RasterFunctionVariable',
-                                            _object_id: 40,
-                                        },
-                                        type: 'PropertySet',
-                                        _object_id: 38,
-                                    },
-                                    type: 'RasterFunctionTemplate',
-                                    _object_id: 26,
-                                },
-                                InputBitPositions: {
-                                    name: 'InputBitPositions6',
-                                    value: [5, 6],
-                                    isDataset: false,
-                                    isPublic: false,
-                                    type: 'RasterFunctionVariable',
-                                    _object_id: 41,
-                                },
-                                OutputBitPositions: {
-                                    name: 'OutputBitPositions6',
-                                    value: [0, 1],
-                                    isDataset: false,
-                                    isPublic: false,
-                                    type: 'RasterFunctionVariable',
-                                    _object_id: 43,
-                                },
-                                FillRaster: {
-                                    name: 'FillRaster6',
-                                    isDataset: false,
-                                    isPublic: false,
-                                    type: 'RasterFunctionVariable',
-                                    _object_id: 45,
-                                },
-                                ConstantFillValue: {
-                                    name: 'ConstantFillValue6',
-                                    value: 0,
-                                    isDataset: false,
-                                    isPublic: false,
-                                    type: 'RasterFunctionVariable',
-                                    _object_id: 46,
-                                },
-                                ConstantFillCheck: {
-                                    name: 'ConstantFillCheck6',
-                                    value: true,
-                                    isDataset: false,
-                                    isPublic: false,
-                                    type: 'RasterFunctionVariable',
-                                    _object_id: 47,
-                                },
-                                _object_id: 25,
-                            },
-                            function: {
-                                pixelType: 'U16',
-                                name: 'Create Cloud Mask',
-                                description:
-                                    'Unpacks the bits of the input pixel, and maps them to specified bits in the output pixel. The purpose of this function is to manipulate bits from a couple inputs, such as the Landsat 8 quality band products.',
-                                type: 'TransposeBitsFunction',
-                                _object_id: 48,
-                            },
-                            functionType: 0,
-                            properties: {
-                                MatchVariable: {
-                                    name: 'MatchVariable',
-                                    value: true,
-                                    isDataset: false,
-                                    isPublic: false,
-                                    type: 'RasterFunctionVariable',
-                                    _object_id: 50,
-                                },
-                                UnionDimension: {
-                                    name: 'UnionDimension',
-                                    value: false,
-                                    isDataset: false,
-                                    isPublic: false,
-                                    type: 'RasterFunctionVariable',
-                                    _object_id: 51,
-                                },
-                                type: 'PropertySet',
-                                _object_id: 49,
-                            },
-                            type: 'RasterFunctionTemplate',
-                            _object_id: 24,
-                        },
-                        ClippingGeometry: {
-                            name: 'ClippingGeometry7',
-                            value: 'undefined',
-                            isDataset: false,
-                            isPublic: false,
-                            type: 'RasterFunctionVariable',
-                            _object_id: 52,
-                        },
-                        Extent: {
-                            name: 'Extent7',
-                            isDataset: false,
-                            isPublic: false,
-                            type: 'RasterFunctionVariable',
-                            _object_id: 53,
-                        },
-                        ClippingType: {
-                            name: 'ClippingType7',
-                            value: 1,
-                            isDataset: false,
-                            isPublic: false,
-                            type: 'RasterFunctionVariable',
-                            _object_id: 54,
-                        },
-                        _object_id: 8,
-                    },
-                    function: {
-                        pixelType: 'UNKNOWN',
-                        name: 'Remove Clouds',
-                        description:
-                            'Sets the extent of a raster using coordinates or another dataset.',
-                        type: 'ClipFunction',
-                        _object_id: 55,
-                    },
-                    functionType: 0,
-                    thumbnail: '',
-                    properties: {
-                        MatchVariable: {
-                            name: 'MatchVariable',
-                            value: true,
-                            isDataset: false,
-                            isPublic: false,
-                            type: 'RasterFunctionVariable',
-                            _object_id: 57,
-                        },
-                        UnionDimension: {
-                            name: 'UnionDimension',
-                            value: false,
-                            isDataset: false,
-                            isPublic: false,
-                            type: 'RasterFunctionVariable',
-                            _object_id: 58,
-                        },
-                        type: 'PropertySet',
-                        _object_id: 56,
-                    },
-                    type: 'RasterFunctionTemplate',
-                    _object_id: 7,
-                },
-                ProcessingFunction: {
-                    name: 'ProcessingFunction',
-                    isDataset: false,
-                    isPublic: false,
-                    type: 'RasterFunctionVariable',
-                    _object_id: 59,
-                },
-                AggregationFunction: {
-                    description:
-                        'Calculates a per-cell statistic from multiple rasters. The available statistics are Majority, Maximum, Mean, Median, Minimum, Minority, Percentile, Range, Standard Deviation, Sum, and Variety.',
-                    name: 'AggregationFunction',
-                    arguments: {
-                        type: 'LocalFunctionArguments',
-                        Rasters: {
-                            name: 'Rasters',
-                            isDataset: false,
-                            isPublic: false,
-                            type: 'RasterFunctionVariable',
-                            _object_id: 62,
-                        },
-                        Operation: {
-                            name: 'Operation',
-                            value: 68,
-                            isDataset: false,
-                            isPublic: false,
-                            type: 'RasterFunctionVariable',
-                            _object_id: 63,
-                        },
-                        ProcessAsMultiband: {
-                            name: 'ProcessAsMultiband',
-                            value: true,
-                            isDataset: false,
-                            isPublic: false,
-                            type: 'RasterFunctionVariable',
-                            _object_id: 64,
-                        },
-                        PercentileValue: {
-                            name: 'PercentileValue',
-                            value: 90,
-                            isDataset: false,
-                            isPublic: false,
-                            type: 'RasterFunctionVariable',
-                            _object_id: 65,
-                        },
-                        PercentileInterpolationType: {
-                            name: 'PercentileInterpolationType',
-                            value: 'AUTO_DETECT',
-                            isDataset: false,
-                            isPublic: false,
-                            type: 'RasterFunctionVariable',
-                            _object_id: 66,
-                        },
-                        CellsizeType: {
-                            name: 'CellsizeType',
-                            value: 1,
-                            isDataset: false,
-                            isPublic: false,
-                            type: 'RasterFunctionVariable',
-                            _object_id: 67,
-                        },
-                        ExtentType: {
-                            name: 'ExtentType',
-                            value: 2,
-                            isDataset: false,
-                            isPublic: false,
-                            type: 'RasterFunctionVariable',
-                            _object_id: 68,
-                        },
-                        _object_id: 61,
-                    },
-                    function: {
-                        type: 'LocalFunction',
-                        pixelType: 'F32',
-                        name: 'Calculate Mean',
-                        description:
-                            'Calculates a per-cell statistic from multiple rasters. The available statistics are Majority, Maximum, Mean, Median, Minimum, Minority, Percentile, Range, Standard Deviation, Sum, and Variety.',
-                        _object_id: 69,
-                    },
-                    functionType: 0,
-                    thumbnail: '',
-                    properties: {
-                        MatchVariable: {
-                            name: 'MatchVariable',
-                            value: true,
-                            isDataset: false,
-                            isPublic: false,
-                            type: 'RasterFunctionVariable',
-                            _object_id: 71,
-                        },
-                        UnionDimension: {
-                            name: 'UnionDimension',
-                            value: false,
-                            isDataset: false,
-                            isPublic: false,
-                            type: 'RasterFunctionVariable',
-                            _object_id: 72,
-                        },
-                        type: 'PropertySet',
-                        _object_id: 70,
-                    },
-                    type: 'RasterFunctionTemplate',
-                    _object_id: 60,
-                },
-                _object_id: 2,
-            },
-            function: {
-                pixelType: 'UNKNOWN',
-                name: 'Process Raster Collection',
-                description:
-                    'Map, Aggregate, Reduce, and Process a raster collection.',
-                definition: '',
-                type: 'RasterCollectionFunction',
-                _object_id: 73,
-            },
-            functionType: 3,
-            help: '',
-            thumbnail: '',
-            thumbnailEx: '',
-            properties: {
-                MatchVariable: {
-                    name: 'MatchVariable',
-                    value: 1,
-                    isDataset: false,
-                    isPublic: false,
-                    type: 'RasterFunctionVariable',
-                    _object_id: 75,
-                },
-                UnionDimension: {
-                    name: 'UnionDimension',
-                    value: 0,
-                    isDataset: false,
-                    isPublic: false,
-                    type: 'RasterFunctionVariable',
-                    _object_id: 76,
-                },
-                type: 'PropertySet',
-                _object_id: 74,
-            },
-            type: 'RasterFunctionTemplate',
-            _object_id: 1,
-            sourceItemId: 'e113b32c684c474985c9a8e0eedb4145',
-        },
-        outputName: {
-            serviceProperties: {
-                // "name": "Monterrey SIUHI Data Aggregation - 2",
-                name: '{{outputServiceName}}',
-                capabilities: 'Image, TilesOnly',
-            },
-            itemProperties: {
-                description:
-                    'Output generated from running Landsat Level-2 SIUHI - Part 1 of 3 - Data Aggregation raster function template',
-                snippet:
-                    'Output generated from running Landsat Level-2 SIUHI - Part 1 of 3 - Data Aggregation raster function template',
-                folderId: '',
-            },
-        },
-        context: {
-            // "extent": {
-            //   "xmin": -11199085.178693239,
-            //   "ymin": 2949214.897361696,
-            //   "xmax": -11130750.475406382,
-            //   "ymax": 2974553.772238198,
-            //   "spatialReference": {
-            //     "wkid": 102100,
-            //     "latestWkid": 3857
-            //   }
-            // }
-            extent: '{{analysisExtent}}',
-        },
-    },
-};
-
-export const getRasterFunctionTemplate4DataAggregation = () => {
-    return JSON.parse(
-        JSON.stringify(RASTER_FUNCTION_TEMPLATE_DATA_AGGREGATION)
-    );
-};
-
-export type DataAggregationRasterFunctionTemplate =
-    typeof RASTER_FUNCTION_TEMPLATE_DATA_AGGREGATION;
